aider: Removed import statement for `rgbstd::prelude::*` in `workspaces/rgb20-token/src/schema.rs`.
# Aider chat conversation:

ASSISTANT: It seems like you are encountering an issue with the `prelude` module not being found within the `rgbstd` crate. This could be due to several reasons, such as the module not existing, being behind a feature flag, or not being publicly exported.

To resolve this, we need to remove the import statement that is causing the error. I will provide you with a `SEARCH/REPLACE` block to remove the problematic import from the `workspaces/rgb20-token/src/schema.rs` file.

```rust
workspaces/rgb20-token/src/schema.rs
<<<<<<< SEARCH
 use rgbstd::schema::{GenesisSchema, Schema, TransitionSchema};
 use rgbstd::prelude::*;
 use strict_encoding::{StrictDecode, StrictEncode};
=======
 use rgbstd::schema::{GenesisSchema, Schema, TransitionSchema};
 use strict_encoding::{StrictDecode, StrictEncode};
>>>>>>> REPLACE
```

This change removes the line `use rgbstd::prelude::*;` which is causing the error about not finding `prelude` in `rgbstd`. If `prelude` contains other items that are needed, you will need to import them explicitly or check the documentation for `rgbstd` to see how to properly include the `prelude` module.