aider: Created User class to manage user data (public and private).
# Aider chat conversation:

USER: > I am creating a Dating DApp with other students in my class. I am part of the Back End team and the provided files are w
hat the class has done so far. The teams are UI, Backend, QA, Immutable services, and Design. The following tasks have to be completed for the project. We are using prismaorm and planet scale.

Manage data (private and public)

Dummy data
No defined architecture

Matching algorithm
Prismaorm
Planet scale

Basic planet scale in prism form with profile info

Seperate subdirectory in back end

Integrate backend scaffolding
Learn Prisma
Learn Planet scale

Get profile data from UI

Define Daemon to make matches

ASSISTANT: It seems like you have a lot of tasks to complete for your project. However, I need more specific instructions to help you modify the code. Could you please provide more details about the changes you want to make in the provided files? For example, do you want to add new functions, modify existing ones, or refactor the code?

USER: I want you to make new files to complete these tasks, dont alter the files provided

ASSISTANT: Understood. However, to provide the most accurate assistance, I need more specific details about each task. For example, for the task "Manage data (private and public)", what kind of data are we talking about? What format is it in? How should it be managed?

For the task "Matching algorithm", what are the criteria for a match?

For the task "Basic planet scale in prism form with profile info", what does a profile contain?

For the task "Get profile data from UI", what kind of data are we expecting from the UI?

For the task "Define Daemon to make matches", what is the matching logic?

Please provide as much detail as possible for each task.

USER: For "Manage data (private and public)" this is the data of each user using the dating app. Since this app is decentralized, we need each user to have some data which is public (such as name and birthday) and some which is private (such as messages between users).

ASSISTANT: Understood. To manage user data, we can create a new file named `user.js` in the `Dating-DApp/svelte-app/src/` directory. This file will contain a `User` class with methods to handle user data.

Here is the *SEARCH/REPLACE block* for the new file:

Dating-DApp/svelte-app/src/user.js
```javascript
<<<<<<< SEARCH
=======
class User {
  constructor(name, birthday, messages) {
    this.publicData = {
      name: name,
      birthday: birthday
    };
    this.privateData = {
      messages: messages
    };
  }

  getPublicData() {
    return this.publicData;
  }

  getPrivateData() {
    return this.privateData;
  }

  addMessage(message) {
    this.privateData.messages.push(message);
  }
}

export default User;
>>>>>>> REPLACE
```

This `User` class has a constructor that takes a name, a birthday, and an array of messages. The name and birthday are stored in a `publicData` object, and the messages are stored in a `privateData` object. The `getPublicData` and `getPrivateData` methods return the respective data objects, and the `addMessage` method adds a new message to the `privateData` object.

Please note that this is a basic implementation and might need to be adjusted based on your specific requirements and the overall architecture of your application.