aider: Add a version tracker file and import it into the index file to display the current version.
# Aider chat conversation:

USER: Add a build artifact to this file so that on each new deployment the version increments.

ASSISTANT: To implement an automatic version increment on each new deployment, we need to add a build artifact that will track the version number. This typically involves creating a script or using a build tool that updates the version number in a file. However, since we are working with a TypeScript file, we can't execute scripts directly from it.

Instead, we can create a new file that will act as a version tracker. This file can be updated manually or by a build script outside of the scope of this TypeScript file. For the purpose of this example, I will show you how to create a new file that will hold the version number and how to import it into your `index.ts`.

Here are the *SEARCH/REPLACE* blocks:

Dating-DApp/final-project/src/lib/version.ts
```typescript
<<<<<<< SEARCH
=======
// This file acts as a version tracker.
// Update this version manually or through an automated script during deployment.
export const version = '1.0.0';
>>>>>>> REPLACE
```

Dating-DApp/final-project/src/lib/index.ts
```typescript
<<<<<<< SEARCH
// place files you want to import through the `$lib` alias in this folder.
=======
// place files you want to import through the `$lib` alias in this folder.
import { version } from './version';

console.log(`Current version: ${version}`);
>>>>>>> REPLACE
```

Please note that the actual incrementing of the version number should be handled by a deployment script or a build tool. The `version.ts` file is just a placeholder for the current version number, and you would need to implement a mechanism to update this file as part of your deployment process.