aider: Add a script to the package.json to increment the version, called update-version.
# Aider chat conversation:

USER: add a script to the package.json to increment the version. Call it update-version

ASSISTANT: To add a script to `package.json` that increments the version, we will use the `npm version` command which can update the version in your `package.json` file and also commit the change to git. Here's how to add the `update-version` script:

Dating-DApp/final-project/package.json
```json
<<<<<<< SEARCH
	"format": "prettier --write ."
=======
	"format": "prettier --write .",
	"update-version": "npm version patch"
>>>>>>> REPLACE
```

This script will increment the patch version (the third number in the version). If you want to increment the minor or major version, you can replace `patch` with `minor` or `major` respectively. After running `npm run update-version`, the version in `package.json` will be incremented, and a new commit with the version change will be created.